[fix] 세션 유저 변경 및 암호화 관련 버그 수정 (#208)
* fix: 세션 null pointer 수정

* fix: 로그인 Social ID 예외 로직 해쉬 ID로 확인하도록 변경

* fix: 탈퇴 시 세션 userId -> user로 가져오도록 변경

* fix: 객체 deep copy로 복호화가 저장되지 않도록 변경

* fix: Get을 위한 decrypte deep copy 메소드와 상태 변경을 위한 조회 메서드 분리

+++ src/test/java/upbrella/be/user/service/UserServiceDynamicTest.java
@@ -53,7 +53,7 @@ class UserServiceDynamicTest {
 
         return List.of(
                 DynamicTest.dynamicTest("새로 가입한 유저는 DB에 저장된다.", () -> {
-                    SessionUser joined = userService.join(user.getSocialId().hashCode(), joinRequest);
+                    SessionUser joined = userService.join((long) user.getSocialId().hashCode(), joinRequest);
                     Optional<User> foundUser = userRepository.findById(joined.getId());
 
                     assertAll(() -> assertTrue(foundUser.isPresent()),
